+++ checker/src/main/java/org/checkerframework/checker/lock/LockAnnotatedTypeFactory.java
@@ -60,6 +60,7 @@ import org.checkerframework.framework.util.dependenttypes.DependentTypesError;
 import org.checkerframework.framework.util.dependenttypes.DependentTypesHelper;
 import org.checkerframework.javacutil.AnnotationBuilder;
 import org.checkerframework.javacutil.AnnotationUtils;
+import org.checkerframework.javacutil.BugInCF;
 import org.checkerframework.javacutil.ElementUtils;
 import org.checkerframework.javacutil.Pair;
 import org.checkerframework.javacutil.TreeUtils;
@@ -94,7 +95,7 @@ public class LockAnnotatedTypeFactory
   /** The @{@link GuardedByUnknown} annotation. */
   protected final AnnotationMirror GUARDEDBYUNKNOWN =
       AnnotationBuilder.fromClass(elements, GuardedByUnknown.class);
-  /** The @{@link GuardedByBottom} annotation. */
+  /** The @{@link GuardedBy} annotation. */
   protected final AnnotationMirror GUARDEDBY =
       createGuardedByAnnotationMirror(new ArrayList<String>());
   /** The @{@link GuardedByBottom} annotation. */
@@ -272,14 +273,14 @@ public class LockAnnotatedTypeFactory
   /** LockQualifierHierarchy. */
   class LockQualifierHierarchy extends MostlyNoElementQualifierHierarchy {
 
+    /** Qualifier kind for the @{@link GuardedByUnknown} annotation. */
+    private final QualifierKind GUARDEDBYUNKNOWN_KIND;
     /** Qualifier kind for the @{@link GuardedBy} annotation. */
     private final QualifierKind GUARDEDBY_KIND;
     /** Qualifier kind for the @{@link GuardSatisfied} annotation. */
     private final QualifierKind GUARDSATISFIED_KIND;
     /** Qualifier kind for the @{@link GuardedByBottom} annotation. */
     private final QualifierKind GUARDEDBYBOTTOM_KIND;
-    /** Qualifier kind for the @{@link GuardedByUnknown} annotation. */
-    private final QualifierKind GUARDEDBYUNKNOWN_KIND;
 
     /**
      * Creates a LockQualifierHierarchy.
@@ -290,10 +291,10 @@ public class LockAnnotatedTypeFactory
     public LockQualifierHierarchy(
         Collection<Class<? extends Annotation>> qualifierClasses, Elements elements) {
       super(qualifierClasses, elements);
+      GUARDEDBYUNKNOWN_KIND = getQualifierKind(GUARDEDBYUNKNOWN);
       GUARDEDBY_KIND = getQualifierKind(GUARDEDBY);
       GUARDSATISFIED_KIND = getQualifierKind(GUARDSATISFIED);
       GUARDEDBYBOTTOM_KIND = getQualifierKind(GUARDEDBYBOTTOM);
-      GUARDEDBYUNKNOWN_KIND = getQualifierKind(GUARDEDBYUNKNOWN);
     }
 
     @Override
@@ -346,7 +347,9 @@ public class LockAnnotatedTypeFactory
       } else if (qualifierKind2 == GUARDEDBYBOTTOM_KIND) {
         return a1;
       }
-      throw new RuntimeException("Unexpected");
+      throw new BugInCF(
+          "leastUpperBoundWithElements(%s, %s, %s, %s, %s)",
+          a1, qualifierKind1, a2, qualifierKind2, lubKind);
     }
 
     @Override
@@ -379,7 +382,9 @@ public class LockAnnotatedTypeFactory
       } else if (qualifierKind2 == GUARDEDBYUNKNOWN_KIND) {
         return a1;
       }
-      throw new RuntimeException("Unexpected");
+      throw new BugInCF(
+          "greatestLowerBoundWithElements(%s, %s, %s, %s, %s)",
+          a1, qualifierKind1, a2, qualifierKind2, glbKind);
     }
   }
 
@@ -605,13 +610,13 @@ public class LockAnnotatedTypeFactory
 
     List<? extends ExpressionTree> methodInvocationTreeArguments =
         ((MethodInvocationTree) tree).getArguments();
-    List<AnnotatedTypeMirror> requiredArgs =
-        AnnotatedTypes.expandVarArgs(this, invokedMethod, methodInvocationTreeArguments);
+    List<AnnotatedTypeMirror> paramTypes =
+        AnnotatedTypes.expandVarArgsParameters(this, invokedMethod, methodInvocationTreeArguments);
 
-    for (int i = 0; i < requiredArgs.size(); i++) {
+    for (int i = 0; i < paramTypes.size(); i++) {
       if (replaceAnnotationInGuardedByHierarchyIfGuardSatisfiedIndexMatches(
           methodDefinitionReturn,
-          requiredArgs.get(i),
+          paramTypes.get(i),
           returnGuardSatisfiedIndex,
           getAnnotatedType(methodInvocationTreeArguments.get(i))
               .getEffectiveAnnotationInHierarchy(GUARDEDBYUNKNOWN))) {