+++ checker/src/test/java/org/checkerframework/checker/test/junit/ainferrunners/AinferTestCheckerJaifsValidationTest.java
@@ -26,7 +26,7 @@ public class AinferTestCheckerJaifsValidationTest extends CheckerFrameworkPerDir
   @Override
   public void run() {
     // Only run if annotated files have been created.
-    // See ainferTests task.
+    // See ainferTest task.
     if (!new File("tests/ainfer-testchecker/annotated/").exists()) {
       throw new RuntimeException(
           AinferTestCheckerJaifsTest.class + " must be run before this test.");