+++ framework/src/main/java/org/checkerframework/framework/flow/CFAbstractTransfer.java
@@ -836,7 +836,8 @@ public abstract class CFAbstractTransfer<
   /**
    * Takes a node, and either returns the node itself again (as a singleton list), or if the node is
    * an assignment node, returns the lhs and rhs (where splitAssignments is applied recursively to
-   * the rhs -- that is, the rhs may not appear in the result, but rather its lhs and rhs may).
+   * the rhs -- that is, it is possible that the rhs does not appear in the result, but rather its
+   * lhs and rhs do).
    *
    * @param node possibly an assignment node
    * @return a list containing all the right- and left-hand sides in the given assignment node; it